Made changes for the satellite data to store the minx, miny values as floats instead of doubles because the psql database tables weren't able to match unique floats, so duplicate entries were getting made, not allowing some data to then be processed.

+++ edexOsgi/com.raytheon.uf.edex.plugin.goesr/src/com/raytheon/uf/edex/plugin/goesr/geospatial/GoesrProjectionFactory.java
@@ -168,8 +168,8 @@ public class GoesrProjectionFactory {
                     coverage.setCrsWKT(crs.toWKT());
                     coverage.setDx(envelope.getDx());
                     coverage.setDy(envelope.getDy());
-                    coverage.setMinX(envelope.getMinX());
-                    coverage.setMinY(envelope.getMinY());
+                    coverage.setMinX((float)envelope.getMinX());
+                    coverage.setMinY((float)envelope.getMinY());
                     coverage.setNx(envelope.getNx());
                     coverage.setNy(envelope.getNy());
                     /* Trigger generation of a location. */
